chore: Add Sitemap to robots.txt
Add a Sitemap directive to the robots.txt file to improve search engine indexing and crawling. The Sitemap URL is set to https://vitaway.org/sitemap.xml.

Update Blog model

Added the 'slug' column to the fillable array in the Blog model. This allows for the storage and retrieval of slugs for blog posts.

Update Generals helper

Added a new helper function 'authUser' to retrieve the authenticated user using the Auth facade.

Update update.blade.php

Changed the 'blog_id' attribute to 'blog_slug' in the update.blade.php file to use the blog slug instead of the blog ID.

Update web.php

Changed the route parameter for the 'single.blog' route in web.php to use the 'slug' column of the 'blogs' table instead of the 'id' column.

Update composer.json

Added the 'spatie/laravel-sitemap' package to the composer.json file as a new dependency.

Create migration for 'slug' column

Created a new migration file to add the 'slug' column to the 'blogs' table.

Update single-blog.blade.php

Changed the URL in the Open Graph and Twitter meta tags in single-blog.blade.php to use the blog slug instead of the blog ID.

Update BlogController

Modified the 'store' method in the BlogController to generate a unique slug for each new blog post using the 'Str' helper class.

Create AsyncPreviousBlogsSlug command

Created a new command 'AsyncPreviousBlogsSlug' to asynchronously generate slugs for all existing blog posts.

Update blogs.blade.php

Changed the URL in the thumbnail and heading links in blogs.blade.php to use the blog slug instead of the blog ID.